New Ship: Woeful Cthonian-class Patrol Cutter (#3624)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
<details><summary>Screenshots</summary>
<p>

![StrongDMM-2024-10-22 21 48
24](https://github.com/user-attachments/assets/4af2be10-6802-4f3a-a91b-bd8f882dbc8e)
![StrongDMM-2024-10-22 21 48
26](https://github.com/user-attachments/assets/7b7c40cd-82af-431e-9ace-9c7dddda2947)

![image](https://github.com/user-attachments/assets/39dc3e77-a58f-4f42-8bd6-1b28d8377a9d)


</p>
</details> 

Adds a new ship to the PGF Roster - the Woeful Cthonian. The Cthonian is
a small ship designed for generalist usage, with a particularly
functional medbay and high levels of style. The crew of a Cthonian is:

- 1 Captain
- 1 Engineer
- 1 Doctor
- 3 Crewmen

You will likely notice that they are all navy - this is to give this
ship a slightly higher focus by taking away the internal dynamic marines
and navy are supposed to have. It also made it way easier to map.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Expands the PGF roster with a small generalist to play in rounds that
don't have the pop nor need for a 10+ person ship
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: A new PGF ship is in the neighborhood - Woeful Cthonians have been
deployed in the frontier!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Erika Fox <94164348+Bjarl@users.noreply.github.com>