Allows medical smocks to use their suit storage. (#3586)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR adds a whitelist of medical items to the suit storage slots of
Frontiersmen, NGR, Hardliners, and Nanotrasen smocks, as well as
expanding the whitelist for the non-factional surgical apron.

Currently, I don't know of any items that work in the smocks aside from
oxytanks and a PDA for the NT smock. Whereas other, more generic
suits(such as winter coats and labcoats) would be more useful to wear by
having a suit slot in them. This is a shame, because the four smock
sprites are very pleasing, and compliment their faction uniform well.
Pulling from labcoats and medical winter coats, with a few tweakings of
my own, I have created a whitelist for the suit storage slot to be used
on each smock, including but not limited to surgical tools, medical
scanners, chemistry tools, and a few generic items.

NOTE: While working on this, I discovered a few things that may warrant
attention. The Hardliner and Frontiersmen smocks are the "base" suit
model for all the factional gear, meaning future edits to them that we
don't want all their equipment to inherit may be difficult if we don't
add a placeholder model. Additionally, in the future, it may be smart to
put the faction AFTER the category of item it is, e.g:
"/clothing/suit/medical_smock/nanotrasen" instead of what it is now,
"/clothing/suit/nanotrasen/medical smock" for easier grouping or
cross-faction changes.

## Why It's Good For The Game
Players will feel less inclined to forgo the cool-looking smocks for
other clothes simply because they don't have suit storage. Makes medical
inventory management slightly easier.

## Changelog

:cl:
add: Added a suit storage whitelist to the Nanotrasen, Hardliner, NGR,
and Frontiersmen smocks. Expands the whitelist of the surgical smock.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->