Changed growth serum so that it is directly proportional to volume in a mob, but it is clamped at 30 units for 30% growth. (#3729)
## About The Pull Request

Title. 30u gives you 30% growth. 50u is still 30%. 10u is 10%.

Also adjusts the descriptions so it's less sexual in nature.

## Why It's Good For The Game

I don't want to remove growth serum. Yeah it doesn't make sense I know,
but let it have a little soul. 30% is noticeable, but not annoying or
intrusive. Here's a 30% grown pori


![dreamseeker_zXXpaYrqnt](https://github.com/user-attachments/assets/b2d37eed-dc2f-4794-9db0-b12f1957245e)

Compare to this round, where I became a T-Rex with Narisi. A bunch of
people laughed and had a good time with the megapori stomping around the
outpost, but now it's time to be reasonable. Admins can always size edit
a pori if they ever wanted such shenanigans again anyways :)


![dreamseeker_6FPaU6POQJ](https://github.com/user-attachments/assets/42f27014-8e16-4fda-ba36-8ab0c9599955)


## Changelog

:cl:
balance: growth serum now maxes out at 30% growth with 30u and has a
smooth sliding scale.
/:cl: