Brighter rail lights (#3722)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Brightens rail lights to be on-par with seclights

## Why It's Good For The Game

Currently as it is, rail lights are about as good as flashlights, and as
a result aren't an attachment most, if any keep for long once they get
other attachments, or in this case a better flashlight. This'll bring
them on-par with seclights, making them a viable choice as an attachment
for longer.

## Changelog

:cl:
balance: Rail lights have been brightened
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->