Ngr+Hardliner digitigrade sprites (#3755)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds ditigrade sprites for the common crew and officers of NGR and
Hardliner factions.

Sprites by @Imaginos16

![Screenshot 2024-11-14
001631](https://github.com/user-attachments/assets/03df5116-499d-4bf4-88f9-15ee7df8cdc2)


## Why It's Good For The Game

Ditigrade sprites have been lacking for both NGR and Hardliner crews.
These sprites by Imaginos adds a handful to the basic and officer
jumpsuits.

## Changelog

:cl: PositiveEntropy
add: Ditigrade sprites for NGR and Hardliner regular and officer
jumpsuits.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->