Crate Shelf Tweaks (#3560)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Mapped crate shelves now start at a height of 2. You can build them to a
max height of 4
Constructed crate shelves start at 1 crate and need to be built up
You can now add shelves to crate shelves (with metal)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
it's not, honestly. i think this change is going to be the death of the
game. for pr after pr i've made nothing but hastily produced shitcode
changes to a game that used to be soulful and thriving, stripping out
the identity of what's made it so interesting to people. this is where
it comes full circle. im sorry for doing this. but it had to be done.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: You can now add shelves to crate racks
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Erika Fox <94164348+Bjarl@users.noreply.github.com>
Co-authored-by: Sun-Soaked <45698967+Sun-Soaked@users.noreply.github.com>