Rock Ruin: N+S Mining Installation (#3534)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
<details><summary>Pictures</summary>
<p>


![dreamseeker_xEj59EPTs5](https://github.com/user-attachments/assets/d92a9fb4-818e-465e-96f0-99a93798eca6)

![dreamseeker_auitVwi8CV](https://github.com/user-attachments/assets/6638cb59-d77f-4000-b6b0-490ba4eeb0b1)

![dreamseeker_scQzFBzfAx](https://github.com/user-attachments/assets/744de12a-4943-458a-87d6-a4f6be7e7e36)
![StrongDMM-2024-10-11 15 15
44](https://github.com/user-attachments/assets/5ba2a741-60c0-4988-9740-fdafcaa14586)
![StrongDMM-2024-10-11 15 15
45](https://github.com/user-attachments/assets/47b5fffd-9faf-41c5-8e99-ac3016ea8f6a)

</p>
</details> 


Adds a new ruin to the rockplanet pool. The N+S Mining Base is a small
mining installation that had been built to extract minerals - and then
was promptly raided by the Ramzi Clique (ramzi guys are somewhat waiting
on #3462. I consider them placeholder right now). The ruin contains a
small amount of weaponry, some fluff, and a refinery for your lovely ore
mining efforts.

The main gimmick is based around cleaning the base up, getting it
somewhat functional, and using the drill to extract ore from the
comically good ore veins. There are platforms with turrets deployed
around them (as well as sandbags) to form a defensible position, but
unless you can get into the refinery, it'll be somewhat annoying to
actually use.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It's gimmicky content!!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: N+S has reported losing contact with one of their rockplanet based
mining installations
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Erika Fox <94164348+Bjarl@users.noreply.github.com>