Fixes the Listening Post Server Monitor + Replaces a Frontier Suit with a Softsuit (#3777)
## About The Pull Request

The Telecomms Monitor board on the listening post doesn't actually work.
It's just a maintenance computer. The intended thing to properly use the
listening setup is a Server Monitor, that lets you see the contents of
the server.

Replaced the singular suit on the ruin with a Soft suit

## Why It's Good For The Game

Fixes good

## Changelog

:cl:
add: Added a soft suit to the Listening Post
fix: Fixed the Listening Post Comms Monitor to be the right type
/:cl: