Modifies snack, drink, and cigarette vender costs (#3721)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This pull request reduces the cost of items in snack, drink, and
cigarette venders.

**Snack Vending Machines**:

![image](https://github.com/user-attachments/assets/44dfff10-8553-4032-a76f-0f625d9a46f6)

**Soda Vending Machines**:

![image](https://github.com/user-attachments/assets/74c25594-2f84-4a34-a335-7eb354d41d4c)

**Tea and Coffee Vending Machines**:

![image](https://github.com/user-attachments/assets/4db2db0b-b45e-4ec2-a6d3-7aa88d1a2250)

**Cigarette Vending Machines**:

![image](https://github.com/user-attachments/assets/1f5cbde6-db22-4171-b12b-608f25376b80)

**Drink Vending Machines**:

|![image](https://github.com/user-attachments/assets/0963fee1-3446-487d-8303-29f22ef4e5bf)|![image](https://github.com/user-attachments/assets/b54ff9db-7221-4914-b6cd-dc844893496a)|![image](https://github.com/user-attachments/assets/4ecec800-2801-4a82-8177-38f3be171195)|
|---|---|---| 

## Why It's Good For The Game

The prices of these items have been brought down from their station-era
level to a price range that seems reasonable for the shiptest universe.

## Changelog

:cl:
balance: reduces drink and cigarette vender item costs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: zimon9 <122945887+zimon9@users.noreply.github.com>